+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.3] - 2023-07-18
+
+### Added
+
+- Added `__convertToFileCURL(...)` function
+
+### Changed
+
+...
+
+### Fixed
+
+- Various bug fixes (in sending InputFile's)
+
+
 ## [1.2.2] - 2023-07-18
 
 ### Added
@@ -15,6 +30,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 ...
+
 
 ## [1.2.0] - 2023-07-18
 
@@ -33,6 +49,8 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+...
+
 ## [1.1.9] - 2023-07-18
 
 ### Added
@@ -48,6 +66,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 ...
+
 
 ## [1.1.8] - 2023-07-16